:lipstick: feat e :recycle: refactor: Adicionando a tela inicial e corrigindo o uso de paths no tsconfig.json
