+++ src/app/content/(attested-sites)/help/[name].tsx
@@ -1,7 +1,7 @@
-import { Text, View } from "react-native";
-import stepsSearch from "@/assets/data/steps-search.json";
+import { ScrollView, Text, View } from "react-native";
 import { useLocalSearchParams } from "expo-router";
-import StepsToSearchCard from "@/components/steps-to-search-card";
+import stepsSearch from "@assets/data/steps-search.json";
+import StepsToSearchCard from "@components/steps-to-search-card";
 
 export default function Help() {
   const { name } = useLocalSearchParams();
@@ -19,8 +19,8 @@ export default function Help() {
   const { steps, tip } = searchResult;
 
   return (
-    <View className="my-4">
+    <ScrollView className="my-4">
       <StepsToSearchCard name={name as string} steps={steps} tip={tip} />
-    </View>
+    </ScrollView>
   );
 }
