+++ src/app/content/source/index.tsx
@@ -1,12 +1,11 @@
-import { Image, ScrollView, StatusBar, Text, View } from "react-native";
+import { Image, ScrollView, Text, View } from "react-native";
 
 export default function Source() {
   return (
     <ScrollView style={{ flex: 1 }} showsVerticalScrollIndicator={false}>
       <View className="w-full px-4 py-8 flex justify-center items-center gap-4">
-        <StatusBar barStyle="light-content" />
         <Image
-          source={require("@/assets/images/Open source-rafiki.png")}
+          source={require("@assets/images/Open source-rafiki.png")}
           style={{ width: "100%", height: 200 }}
         />
         <Text className="text-center text-2xl font-bold text-[#173058]">
