+++ src/styles/index.ts
@@ -1,20 +1,1 @@
-import { StyleSheet } from "react-native";
-
-export const styles = StyleSheet.create({
-  shadow: {
-    shadowOffset: {
-      width: 2,
-      height: 2,
-    },
-    shadowColor: "#333",
-    shadowOpacity: 0.3,
-    shadowRadius: 4,
-    elevation: 5,
-  },
-  icon: {
-    flex: 1,
-  },
-  text: {
-    flex: 2,
-  },
-});
+export * from "./boxShadow";
