+++ tsconfig.json
@@ -2,20 +2,12 @@
   "extends": "expo/tsconfig.base",
   "compilerOptions": {
     "strict": true,
-    "baseUrl": "./src",
+    "baseUrl": "./",
     "paths": {
-      "@assets": [
-        "./src/assets/*"
-      ],
-      "@components": [
-        "./src/components/*"
-      ],
-      "@styles": [
-        "./src/styles/*"
-      ],
-      "@/*": [
-        "./*"
-      ]
+      "@components/*": ["./src/components/*"],
+      "@styles/*": ["./src/styles/*"],
+      "@assets/*": ["./assets/*"],
+      "@/*": ["./src/*"]
     }
   },
   "include": [
